исправил ошибки
