+++ project/src/components/header/header.tsx
@@ -1,27 +1,42 @@
-import { Link } from 'react-router-dom';
+import { MouseEventHandler } from 'react';
+import { Link, useNavigate } from 'react-router-dom';
+import { AppRoute } from '../../const';
+import { Offer } from '../../types/offer';
 
-function Header(): JSX.Element {
+type HeaderProps = {
+  offers: Offer[];
+}
+
+function Header({offers}: HeaderProps): JSX.Element {
+  const navigate = useNavigate();
+  const handleGoMainClick: MouseEventHandler<HTMLAnchorElement> = (evt) => {
+    evt.preventDefault();
+    navigate(AppRoute.Main);
+  };
+
+  const favoriteOffers = offers.filter((offer) => offer.isFavorite);
+
   return (
     <header className="header">
       <div className="container">
         <div className="header__wrapper">
           <div className="header__left">
-            <Link className="header__logo-link" to="/">
-              <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41" />
+            <Link className="header__logo-link" to="#" onClick={handleGoMainClick}>
+              <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
             </Link>
           </div>
           <nav className="header__nav">
             <ul className="header__nav-list">
               <li className="header__nav-item user">
-                <Link className="header__nav-link header__nav-link--profile" to="/">
+                <Link className="header__nav-link header__nav-link--profile" to="#">
                   <div className="header__avatar-wrapper user__avatar-wrapper">
                   </div>
                   <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                  <span className="header__favorite-count">3</span>
+                  <span className="header__favorite-count">{favoriteOffers.length}</span>
                 </Link>
               </li>
               <li className="header__nav-item">
-                <Link className="header__nav-link" to="/">
+                <Link className="header__nav-link" to="#">
                   <span className="header__signout">Sign out</span>
                 </Link>
               </li>
