+++ project/src/pages/favorites-screen/favorites-screen.tsx
@@ -1,7 +1,6 @@
-import { MouseEventHandler } from 'react';
-import { Link, useNavigate } from 'react-router-dom';
+import { Link} from 'react-router-dom';
+import Header from '../../components/header/header';
 import Rating from '../../components/rating/rating';
-import { AppRoute } from '../../const';
 import { Offer } from '../../types/offer';
 type FavoritesScreenProps = {
   offers: Offer[];
@@ -9,42 +8,10 @@ type FavoritesScreenProps = {
 function FavoritesScreen({offers}: FavoritesScreenProps) :JSX.Element {
   const favoriteOffers = offers.filter((offer) => offer.isFavorite);
   const cities = Array.from(new Set(favoriteOffers.map((offer) => offer.city.name)));
-  const navigate = useNavigate();
-  const handleGoMainClick: MouseEventHandler<HTMLAnchorElement> = (evt) => {
-    evt.preventDefault();
-    navigate(AppRoute.Main);
-  };
 
   return(
     <div className="page">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link" to="#" onClick={handleGoMainClick}>
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link className="header__nav-link header__nav-link--profile" to="#">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                    <span className="header__favorite-count">{favoriteOffers.length}</span>
-                  </Link>
-                </li>
-                <li className="header__nav-item">
-                  <Link className="header__nav-link" to="#">
-                    <span className="header__signout">Sign out</span>
-                  </Link>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
+      <Header offers={offers}/>
 
       <main className="page__main page__main--favorites">
         <div className="page__favorites-container container">
@@ -61,7 +28,7 @@ function FavoritesScreen({offers}: FavoritesScreenProps) :JSX.Element {
                     </div>
                   </div>
                   <div className="favorites__places">
-                    {favoriteOffers
+                    {offers.filter((offer) => offer.isFavorite)
                       .filter(({ city: { name } }) => name === city)
                       .map(({ id, isPremium, previewImage, price, isFavorite, rating, title, type }) => (
                         <article key={`${city}-${id}`} className="favorites__card place-card">
