+++ project/src/pages/main-screen/main-screen.tsx
@@ -1,4 +1,5 @@
 import { Link } from 'react-router-dom';
+import Header from '../../components/header/header';
 import Offers from '../../components/offers/offers';
 import { Offer } from '../../types/offer';
 
@@ -6,39 +7,10 @@ type MainScreenProps = {
   offers: Offer[];
 }
 function MainScreen({offers} : MainScreenProps) : JSX.Element {
-  const favoriteOffers = offers.filter((offer) => offer.isFavorite);
 
   return(
     <div className="page page--gray page--main">
-      <header className="header">
-        <div className="container">
-          <div className="header__wrapper">
-            <div className="header__left">
-              <Link className="header__logo-link header__logo-link--active" to="/">
-                <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
-              </Link>
-            </div>
-            <nav className="header__nav">
-              <ul className="header__nav-list">
-                <li className="header__nav-item user">
-                  <Link className="header__nav-link header__nav-link--profile" to="/favorites">
-                    <div className="header__avatar-wrapper user__avatar-wrapper">
-                    </div>
-                    <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                    <span className="header__favorite-count">{favoriteOffers.length}</span>
-                  </Link>
-                </li>
-                <li className="header__nav-item">
-                  <Link className="header__nav-link" to="#">
-                    <span className="header__signout">Sign out</span>
-                  </Link>
-                </li>
-              </ul>
-            </nav>
-          </div>
-        </div>
-      </header>
-
+      <Header offers={offers}/>
       <main className="page__main page__main--index">
         <h1 className="visually-hidden">Cities</h1>
         <div className="tabs">
